steering from float to int

+++ ROS/src/control/ugr_ros_control/pegasus_control/include/pegasus_control/cantools.h
@@ -78,7 +78,7 @@ struct odrive_set_input_steering_t {
      * Scale: 1
      * Offset: 0
      */
-    float input_steering;
+    int input_steering;
 };
 
 /**
